ADD freespin info to getgames service API request. FIX currency rounding problem

+++ src/main/java/com/dashur/integration/extw/connectors/relaxgaming/data/service/GameInfo.java
@@ -33,7 +33,7 @@ public class GameInfo {
   private FreeRoundsInfo freespins;
 
   @JsonProperty("legalbetsizes")
-  private List<Integer> legalBetSizes;  // In euro cents
+  private List<Long> legalBetSizes;  // In euro cents
 
   // optional
